Colocalisation testing using regression coefficients

Colocalisation testing supplying only regression coefficients and their variance-covariants matrices

coloc.test.summary(b1, b2, V1, V2, k = 1, plot.coeff = FALSE,
  plots.extra = NULL, bayes = !is.null(bayes.factor), n.approx = 1001,
  level.ci = 0.95, bayes.factor = NULL, bma = FALSE)

Arguments

b1
regression coefficients for trait 1

b2
regression coefficients for trait 2

V1
variance-covariance matrix for trait 1

V2
variance-covariance matrix for trait 2

k
Theta has a Cauchy(0,k) prior. The default, k=1, is equivalent to a uniform (uninformative) prior. We have found varying k to have little effect on the results.

plot.coeff
DEPRECATED. Please plot() returned object instead. TRUE if you want to generate a plot showing the coefficients from the two regressions together with confidence regions.

plots.extra
list with 2 named elements, x and y, equal length character vectors containing the names of the quantities to be plotted on the x and y axes.

x is generally a sequence of theta and eta, with y selected from post.theta, the posterior density of theta, chisq, the chi-square values of the test, and lhood, the likelihood function.

bayes
Logical, indicating whether to perform Bayesian inference for the coefficient of proportionality, eta. If bayes.factor is supplied, Bayes factors are additionally computed for the specificed values. This can add a little time as it requires numerical integration, so can be set to FALSE to save time in simulations, for example.

level.ci, n.approx
level.ci denotes the required level of the credible interval for eta. This is calculated numerically by approximating the posterior distribution at n.approx distinct values.

bayes.factor
Calculate Bayes Factors to compare specific values of eta. bayes.factor should either a numeric vector, giving single value(s) of eta or a list of numeric vectors, each of length two and specifying ranges of eta which should be compared to each other. Thus, the vector or list needs to have length at least two.

bma
parameter set to TRUE when coloc.test is called by coloc.bma. DO NOT SET THIS WHEN CALLING coloc.test DIRECTLY!

Value

an object of class coloc, colocBayes or colocBMA

Details

Typically this should be called from coloc.test() or coloc.bma(), but is left as a public function, to use at your own risk, if you have some other way to define the SNPs under test.