Class "colocPCs"

%% ~~ A concise (1-5 lines) description of what the class is. ~~ Class designed to hold objects returned by function pcs.prepare which generates a principal component summary of two genotype matrices in a form suitable for use in the function pcs.model.

%% ~~ A concise (1-5 lines) description of what the class is. ~~ Class designed to hold objects returned by function pcs.prepare which generates a principal component summary of two genotype matrices in a form suitable for use in the function pcs.model.

Objects from the Class

Objects can be created by calls to the function pcs.prepare(). %% ~~ describe objects here ~~

Objects can be created by calls to the function pcs.prepare(). %% ~~ describe objects here ~~

References

Wallace et al (2012). Statistical colocalisation of monocyte gene expression and genetic risk variants for type 1 diabetes. Hum Mol Genet 21:2815-2824. http://europepmc.org/abstract/MED/22403184

Plagnol et al (2009). Statistical independence of the colocalized association signals for type 1 diabetes and RPS26 gene expression on chromosome 12q13. Biostatistics 10:327-34. http://www.ncbi.nlm.nih.gov/pubmed/19039033

Wallace et al (2012). Statistical colocalisation of monocyte gene expression and genetic risk variants for type 1 diabetes. Hum Mol Genet 21:2815-2824. http://europepmc.org/abstract/MED/22403184

Plagnol et al (2009). Statistical independence of the colocalized association signals for type 1 diabetes and RPS26 gene expression on chromosome 12q13. Biostatistics 10:327-34. http://www.ncbi.nlm.nih.gov/pubmed/19039033

See also

Examples

showClass("colocPCs")

#> Class "colocPCs" [package "coloc"]
#> 
#> Slots:
#>                                       
#> Name:      pcs   group     use    vars
#> Class:  matrix numeric logical numeric

showClass("colocPCs")

#> Class "colocPCs" [package "coloc"]
#> 
#> Slots:
#>                                       
#> Name:      pcs   group     use    vars
#> Class:  matrix numeric logical numeric