Simulated data to use in testing and vignettes in the coloc package

Simulated data to use in testing and vignettes in the coloc package

data(coloc_test_data)

Format

A four of two coloc-style datasets. Elements D1 and D2 have a single shared causal variant, and 50 SNPs. Elements D3 and D4 have 100 SNPs, one shared causal variant, and one variant unique to D3. Use these as examples of what a coloc-style dataset for a quantitative trait should look like.

Examples

data(coloc_test_data)
names(coloc_test_data)

#> [1] "D1" "D2" "D3" "D4"

str(coloc_test_data$D1)

#> List of 9
#>  $ beta    : Named num [1:50] 0.02074 0.01515 -0.00745 0.02269 -0.00422 ...
#>   ..- attr(*, "names")= chr [1:50] "s1" "s2" "s3" "s4" ...
#>  $ varbeta : Named num [1:50] 0.00127 0.00126 0.00124 0.00126 0.00122 ...
#>   ..- attr(*, "names")= chr [1:50] "s1" "s2" "s3" "s4" ...
#>  $ N       : int 10000
#>  $ sdY     : num 1.51
#>  $ type    : chr "quant"
#>  $ MAF     : Named num [1:50] 0.0997 0.0975 0.0976 0.0982 0.1005 ...
#>   ..- attr(*, "names")= chr [1:50] "s1" "s2" "s3" "s4" ...
#>  $ LD      : num [1:50, 1:50] 1 0.738 0.634 0.556 0.478 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:50] "s1" "s2" "s3" "s4" ...
#>   .. ..$ : chr [1:50] "s1" "s2" "s3" "s4" ...
#>  $ snp     : chr [1:50] "s1" "s2" "s3" "s4" ...
#>  $ position: int [1:50] 1 2 3 4 5 6 7 8 9 10 ...

check.dataset(coloc_test_data$D1) # should return NULL if data structure is ok

#> NULL