Bayesian finemapping analysis

Bayesian finemapping analysis

finemap.abf(dataset, p1 = 1e-04)

Arguments

dataset
a list with the following elements

pvalues

P-values for each SNP in dataset 1

N

Number of samples in dataset 1

MAF

minor allele frequency of the variants

beta

regression coefficient for each SNP from dataset 1

varbeta

variance of beta

type

the type of data in dataset 1 - either "quant" or "cc" to denote quantitative or case-control

s

for a case control dataset, the proportion of samples in dataset 1 that are cases

sdY

for a quantitative trait, the population standard deviation of the trait. if not given, it can be estimated from the vectors of varbeta and MAF

snp

a character vector of snp ids, optional. If present, it will be used to merge dataset1 and dataset2. Otherwise, the function assumes dataset1 and dataset2 contain results for the same SNPs in the same order.

Some of these items may be missing, but you must give

alwaystype

if type=="cc"s

if type=="quant" and sdY knownsdY

if type=="quant" and sdY unknownbeta, varbeta, N, MAF and then either

pvalues, MAF

beta, varbeta

p1
prior probability a SNP is associated with the trait 1, default 1e-4

Value

a data.frame:

an annotated version of the input data containing log Approximate Bayes Factors and intermediate calculations, and the posterior probability of the SNP being causal

Details

This function calculates posterior probabilities of different causal variant for a single trait.

If regression coefficients and variances are available, it calculates Bayes factors for association at each SNP. If only p values are available, it uses an approximation that depends on the SNP's MAF and ignores any uncertainty in imputation. Regression coefficients should be used if available.