Run susie on a single coloc-structured dataset

run susie_rss storing some additional information for coloc

runsusie(
  d,
  suffix = 1,
  nref = NULL,
  p = NULL,
  trimz = NULL,
  r2.prune = NULL,
  s_init = NULL,
  ...
)

Arguments

d
coloc dataset, must include LD (signed correlation matrix)

suffix
suffix label that will be printed with any error messages

nref
number of samples in the dataset used to estimate the LD matrix in d

p
prior probability a snp is causal (equivalent to p1 or p2 in coloc.abf). By default, this is set to NULL, upon which we will set a small null_weight to pass to susie_rss() (see vignette a06 for details why). You can override this by setting p as you would p1 or p2 in a coloc function, but note that you may miss some true signals that way. Also note that neither of these options correspond to the susie_rss() defaults, because our goal here is not fine mapping alone.

trimz
used to trim datasets for development purposes

r2.prune
sometimes SuSiE can return multiple signals in high LD. if you set r2.prune to a value between 0 and 1, sets with index SNPs with LD greater than r2.prune

s_init
used internally to extend runs that haven't converged. don't use.

...
arguments passed to susie_rss. In particular, if you want to match some coloc defaults, set

prior_variance=0.2^2 (if a case-control trait) or (0.15/sd(Y))^2 if a quantitative trait

estimate_prior_variance=FALSE

otherwise susie_rss will estimate the prior variance itself

Value

results of a susie_rss run, with some added dimnames

Examples

library(coloc)
data(coloc_test_data)
if(requireNamespace("susieR", quietly = TRUE)) {
  result=runsusie(coloc_test_data$D1,nref=500)
  summary(result)
}

#> running iterations: 100

#> 	converged: TRUE

#> 
#> Variables in credible sets:
#> 
#>  variable variable_prob cs
#>        25             1  1
#> 
#> Credible sets summary:
#> 
#>  cs cs_log10bf cs_avg_r2 cs_min_r2 variable
#>   1    50.8625         1         1       25